+++ sos21-database/src/command/insert_user.rs
@@ -15,7 +15,6 @@ where
         kana_last_name,
         email,
         phone_number,
-        affiliation,
         role,
         category,
         assignment,
@@ -35,14 +34,13 @@ INSERT INTO users (
     kana_last_name,
     email,
     phone_number,
-    affiliation,
     role,
     category,
     assignment,
     assignment_owner_project_id,
     assignment_subowner_project_id,
     assignment_owner_pending_project_id
-) VALUES ( $1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15 )
+) VALUES ( $1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14 )
 "#,
         id,
         created_at,
@@ -52,7 +50,6 @@ INSERT INTO users (
         kana_last_name,
         email,
         phone_number,
-        affiliation,
         role as _,
         category as _,
         assignment as _,