+++ sos21-domain/src/model/project_query.rs
@@ -132,11 +132,11 @@ mod tests {
             category: None,
             attributes: ProjectAttributes::from_attributes(vec![]).unwrap(),
         };
-        let project1 = test_model::new_general_project(test_model::new_user_id());
+        let project1 = test_model::new_general_online_project(test_model::new_user_id());
         assert!(conj.check_project(&project1));
         let project2 = test_model::new_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[
                 ProjectAttribute::Artistic,
                 ProjectAttribute::Academic,
@@ -152,11 +152,12 @@ mod tests {
             category: None,
             attributes: ProjectAttributes::from_attributes(vec![]).unwrap(),
         };
-        let pending_project = test_model::new_general_pending_project(test_model::new_user_id());
+        let pending_project =
+            test_model::new_general_online_pending_project(test_model::new_user_id());
         assert!(conj.check_pending_project(&pending_project));
         let pending_project = test_model::new_pending_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[
                 ProjectAttribute::Artistic,
                 ProjectAttribute::Academic,
@@ -175,19 +176,19 @@ mod tests {
         };
         let project1 = test_model::new_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[],
         );
         assert!(!conj.check_project(&project1));
         let project2 = test_model::new_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[ProjectAttribute::Artistic, ProjectAttribute::Committee],
         );
         assert!(conj.check_project(&project2));
         let project3 = test_model::new_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[
                 ProjectAttribute::Artistic,
                 ProjectAttribute::Academic,
@@ -197,7 +198,7 @@ mod tests {
         assert!(conj.check_project(&project3));
         let project4 = test_model::new_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[ProjectAttribute::Academic, ProjectAttribute::Committee],
         );
         assert!(!conj.check_project(&project4));
@@ -212,19 +213,19 @@ mod tests {
         };
         let pending_project = test_model::new_pending_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[],
         );
         assert!(!conj.check_pending_project(&pending_project));
         let pending_project = test_model::new_pending_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[ProjectAttribute::Artistic, ProjectAttribute::Committee],
         );
         assert!(conj.check_pending_project(&pending_project));
         let pending_project = test_model::new_pending_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[
                 ProjectAttribute::Artistic,
                 ProjectAttribute::Academic,
@@ -234,7 +235,7 @@ mod tests {
         assert!(conj.check_pending_project(&pending_project));
         let pending_project = test_model::new_pending_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[ProjectAttribute::Academic, ProjectAttribute::Committee],
         );
         assert!(!conj.check_pending_project(&pending_project));
@@ -252,19 +253,19 @@ mod tests {
         };
         let project1 = test_model::new_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[],
         );
         assert!(!conj.check_project(&project1));
         let project2 = test_model::new_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[ProjectAttribute::Artistic, ProjectAttribute::Committee],
         );
         assert!(conj.check_project(&project2));
         let project3 = test_model::new_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[
                 ProjectAttribute::Artistic,
                 ProjectAttribute::Academic,
@@ -274,7 +275,7 @@ mod tests {
         assert!(conj.check_project(&project3));
         let project4 = test_model::new_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[ProjectAttribute::Artistic, ProjectAttribute::Academic],
         );
         assert!(!conj.check_project(&project4));
@@ -292,19 +293,19 @@ mod tests {
         };
         let pending_project = test_model::new_pending_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[],
         );
         assert!(!conj.check_pending_project(&pending_project));
         let pending_project = test_model::new_pending_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[ProjectAttribute::Artistic, ProjectAttribute::Committee],
         );
         assert!(conj.check_pending_project(&pending_project));
         let pending_project = test_model::new_pending_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[
                 ProjectAttribute::Artistic,
                 ProjectAttribute::Academic,
@@ -314,7 +315,7 @@ mod tests {
         assert!(conj.check_pending_project(&pending_project));
         let pending_project = test_model::new_pending_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[ProjectAttribute::Artistic, ProjectAttribute::Academic],
         );
         assert!(!conj.check_pending_project(&pending_project));
@@ -323,26 +324,30 @@ mod tests {
     #[test]
     fn test_conj_category() {
         let conj = ProjectQueryConjunction {
-            category: Some(ProjectCategory::General),
+            category: Some(ProjectCategory::GeneralOnline),
             attributes: ProjectAttributes::from_attributes(vec![]).unwrap(),
         };
-        assert!(conj.check_project(&test_model::new_general_project(test_model::new_user_id())));
-        assert!(!conj.check_project(&test_model::new_stage_project(test_model::new_user_id())));
+        assert!(conj.check_project(&test_model::new_general_online_project(
+            test_model::new_user_id()
+        )));
+        assert!(!conj.check_project(&test_model::new_stage_online_project(
+            test_model::new_user_id()
+        )));
     }
 
     #[test]
     fn test_pending_project_conj_category() {
         let conj = ProjectQueryConjunction {
-            category: Some(ProjectCategory::General),
+            category: Some(ProjectCategory::GeneralOnline),
             attributes: ProjectAttributes::from_attributes(vec![]).unwrap(),
         };
         assert!(
-            conj.check_pending_project(&test_model::new_general_pending_project(
+            conj.check_pending_project(&test_model::new_general_online_pending_project(
                 test_model::new_user_id()
             ))
         );
         assert!(
-            !conj.check_pending_project(&test_model::new_stage_pending_project(
+            !conj.check_pending_project(&test_model::new_stage_online_pending_project(
                 test_model::new_user_id()
             ))
         );
@@ -351,7 +356,7 @@ mod tests {
     #[test]
     fn test_conj_complex() {
         let conj = ProjectQueryConjunction {
-            category: Some(ProjectCategory::Stage),
+            category: Some(ProjectCategory::StageOnline),
             attributes: ProjectAttributes::from_attributes(vec![
                 ProjectAttribute::Academic,
                 ProjectAttribute::Committee,
@@ -360,25 +365,25 @@ mod tests {
         };
         let project1 = test_model::new_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[],
         );
         assert!(!conj.check_project(&project1));
         let project2 = test_model::new_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[ProjectAttribute::Academic, ProjectAttribute::Committee],
         );
         assert!(!conj.check_project(&project2));
         let project3 = test_model::new_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::Stage,
+            ProjectCategory::StageOnline,
             &[ProjectAttribute::Academic, ProjectAttribute::Committee],
         );
         assert!(conj.check_project(&project3));
         let project4 = test_model::new_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::Stage,
+            ProjectCategory::StageOnline,
             &[ProjectAttribute::Academic],
         );
         assert!(!conj.check_project(&project4));
@@ -387,30 +392,31 @@ mod tests {
     #[test]
     fn test_pending_project_conj_complex() {
         let conj = ProjectQueryConjunction {
-            category: Some(ProjectCategory::Stage),
+            category: Some(ProjectCategory::StageOnline),
             attributes: ProjectAttributes::from_attributes(vec![
                 ProjectAttribute::Academic,
                 ProjectAttribute::Committee,
             ])
             .unwrap(),
         };
-        let pending_project = test_model::new_general_pending_project(test_model::new_user_id());
+        let pending_project =
+            test_model::new_general_online_pending_project(test_model::new_user_id());
         assert!(!conj.check_pending_project(&pending_project));
         let pending_project = test_model::new_pending_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[ProjectAttribute::Academic, ProjectAttribute::Committee],
         );
         assert!(!conj.check_pending_project(&pending_project));
         let pending_project = test_model::new_pending_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::Stage,
+            ProjectCategory::StageOnline,
             &[ProjectAttribute::Academic, ProjectAttribute::Committee],
         );
         assert!(conj.check_pending_project(&pending_project));
         let pending_project = test_model::new_pending_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::Stage,
+            ProjectCategory::StageOnline,
             &[ProjectAttribute::Academic],
         );
         assert!(!conj.check_pending_project(&pending_project));
@@ -425,13 +431,13 @@ mod tests {
         let query = ProjectQuery::from_conjunctions(vec![conj]).unwrap();
         let project1 = test_model::new_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[],
         );
         assert!(query.check_project(&project1));
         let project2 = test_model::new_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[
                 ProjectAttribute::Artistic,
                 ProjectAttribute::Academic,
@@ -448,11 +454,12 @@ mod tests {
             attributes: ProjectAttributes::from_attributes(vec![]).unwrap(),
         };
         let query = ProjectQuery::from_conjunctions(vec![conj]).unwrap();
-        let pending_project = test_model::new_general_pending_project(test_model::new_user_id());
+        let pending_project =
+            test_model::new_general_online_pending_project(test_model::new_user_id());
         assert!(query.check_pending_project(&pending_project));
         let pending_project = test_model::new_pending_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[
                 ProjectAttribute::Artistic,
                 ProjectAttribute::Academic,
@@ -467,13 +474,13 @@ mod tests {
         let query = ProjectQuery::from_conjunctions(vec![]).unwrap();
         let project1 = test_model::new_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[],
         );
         assert!(!query.check_project(&project1));
         let project2 = test_model::new_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[
                 ProjectAttribute::Artistic,
                 ProjectAttribute::Academic,
@@ -486,11 +493,12 @@ mod tests {
     #[test]
     fn test_pending_project_contradiction() {
         let query = ProjectQuery::from_conjunctions(vec![]).unwrap();
-        let pending_project = test_model::new_general_pending_project(test_model::new_user_id());
+        let pending_project =
+            test_model::new_general_online_pending_project(test_model::new_user_id());
         assert!(!query.check_pending_project(&pending_project));
         let pending_project = test_model::new_pending_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[
                 ProjectAttribute::Artistic,
                 ProjectAttribute::Academic,
@@ -503,7 +511,7 @@ mod tests {
     #[test]
     fn test_complex() {
         let conj1 = ProjectQueryConjunction {
-            category: Some(ProjectCategory::Stage),
+            category: Some(ProjectCategory::StageOnline),
             attributes: ProjectAttributes::from_attributes(vec![ProjectAttribute::Academic])
                 .unwrap(),
         };
@@ -518,25 +526,25 @@ mod tests {
         let query = ProjectQuery::from_conjunctions(vec![conj1, conj2]).unwrap();
         let project1 = test_model::new_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[ProjectAttribute::Academic],
         );
         assert!(!query.check_project(&project1));
         let project2 = test_model::new_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::Stage,
+            ProjectCategory::StageOnline,
             &[ProjectAttribute::Academic],
         );
         assert!(query.check_project(&project2));
         let project3 = test_model::new_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::Stage,
+            ProjectCategory::StageOnline,
             &[ProjectAttribute::Academic, ProjectAttribute::Committee],
         );
         assert!(query.check_project(&project3));
         let project4 = test_model::new_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[ProjectAttribute::Academic, ProjectAttribute::Committee],
         );
         assert!(query.check_project(&project4));
@@ -545,7 +553,7 @@ mod tests {
     #[test]
     fn test_pending_project_complex() {
         let conj1 = ProjectQueryConjunction {
-            category: Some(ProjectCategory::Stage),
+            category: Some(ProjectCategory::StageOnline),
             attributes: ProjectAttributes::from_attributes(vec![ProjectAttribute::Academic])
                 .unwrap(),
         };
@@ -560,25 +568,25 @@ mod tests {
         let query = ProjectQuery::from_conjunctions(vec![conj1, conj2]).unwrap();
         let pending_project = test_model::new_pending_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[ProjectAttribute::Academic],
         );
         assert!(!query.check_pending_project(&pending_project));
         let pending_project = test_model::new_pending_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::Stage,
+            ProjectCategory::StageOnline,
             &[ProjectAttribute::Academic],
         );
         assert!(query.check_pending_project(&pending_project));
         let pending_project = test_model::new_pending_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::Stage,
+            ProjectCategory::StageOnline,
             &[ProjectAttribute::Academic, ProjectAttribute::Committee],
         );
         assert!(query.check_pending_project(&pending_project));
         let pending_project = test_model::new_pending_project_with_attributes(
             test_model::new_user_id(),
-            ProjectCategory::General,
+            ProjectCategory::GeneralOnline,
             &[ProjectAttribute::Academic, ProjectAttribute::Committee],
         );
         assert!(query.check_pending_project(&pending_project));
@@ -589,13 +597,13 @@ mod tests {
         use std::collections::HashSet;
 
         let conj1 = ProjectQueryConjunction {
-            category: Some(ProjectCategory::Stage),
+            category: Some(ProjectCategory::StageOnline),
             attributes: ProjectAttributes::from_attributes(vec![ProjectAttribute::Academic])
                 .unwrap(),
         };
         assert_eq!(
             conj1.possible_categories().collect::<Vec<_>>(),
-            vec![ProjectCategory::Stage]
+            vec![ProjectCategory::StageOnline]
         );
         let conj2 = ProjectQueryConjunction {
             category: None,
@@ -613,12 +621,12 @@ mod tests {
         use std::collections::HashSet;
 
         let conj1 = ProjectQueryConjunction {
-            category: Some(ProjectCategory::Stage),
+            category: Some(ProjectCategory::StageOnline),
             attributes: ProjectAttributes::from_attributes(vec![ProjectAttribute::Academic])
                 .unwrap(),
         };
         let conj2 = ProjectQueryConjunction {
-            category: Some(ProjectCategory::Food),
+            category: Some(ProjectCategory::FoodPhysical),
             attributes: ProjectAttributes::from_attributes(vec![
                 ProjectAttribute::Academic,
                 ProjectAttribute::Committee,
@@ -628,14 +636,14 @@ mod tests {
         let query1 = ProjectQuery::from_conjunctions(vec![conj1, conj2]).unwrap();
         assert_eq!(
             query1.possible_categories().collect::<HashSet<_>>(),
-            [ProjectCategory::Stage, ProjectCategory::Food]
+            [ProjectCategory::StageOnline, ProjectCategory::FoodPhysical]
                 .iter()
                 .copied()
                 .collect::<HashSet<_>>()
         );
 
         let conj1 = ProjectQueryConjunction {
-            category: Some(ProjectCategory::Stage),
+            category: Some(ProjectCategory::StageOnline),
             attributes: ProjectAttributes::from_attributes(vec![ProjectAttribute::Academic])
                 .unwrap(),
         };