+++ sos21-use-case/src/export_projects.rs
@@ -52,10 +52,12 @@ pub struct InputFieldNames {
 
 #[derive(Debug, Clone)]
 pub struct InputCategoryNames {
-    pub general: String,
-    pub stage: String,
-    pub cooking: String,
-    pub food: String,
+    pub general_online: String,
+    pub general_physical: String,
+    pub stage_online: String,
+    pub stage_physical: String,
+    pub cooking_physical: String,
+    pub food_physical: String,
 }
 
 #[tracing::instrument(skip(ctx))]
@@ -318,10 +320,12 @@ where
 
     if category.is_some() {
         let category_name = match data.project.category() {
-            project::ProjectCategory::General => &input.category_names.general,
-            project::ProjectCategory::Stage => &input.category_names.stage,
-            project::ProjectCategory::Cooking => &input.category_names.cooking,
-            project::ProjectCategory::Food => &input.category_names.food,
+            project::ProjectCategory::GeneralOnline => &input.category_names.general_online,
+            project::ProjectCategory::GeneralPhysical => &input.category_names.general_physical,
+            project::ProjectCategory::StageOnline => &input.category_names.stage_online,
+            project::ProjectCategory::StagePhysical => &input.category_names.stage_physical,
+            project::ProjectCategory::CookingPhysical => &input.category_names.cooking_physical,
+            project::ProjectCategory::FoodPhysical => &input.category_names.food_physical,
         };
         writer.write_field(category_name)?;
     }
@@ -446,8 +450,8 @@ mod tests {
 
     async fn prepare_app(login_user: domain::user::User) -> Login<test::context::MockApp> {
         let other = test::model::new_general_user();
-        let project1 = test::model::new_general_project(login_user.id().clone());
-        let project2 = test::model::new_general_project(other.id().clone());
+        let project1 = test::model::new_general_online_project(login_user.id().clone());
+        let project2 = test::model::new_general_online_project(other.id().clone());
 
         let app = test::build_mock_app()
             .users(vec![login_user.clone(), other.clone()])
@@ -490,10 +494,12 @@ mod tests {
             attribute_outdoor: Some("屋外企画".to_string()),
         };
         let category_names = export_projects::InputCategoryNames {
-            general: "一般".to_string(),
-            stage: "ステージ".to_string(),
-            cooking: "調理".to_string(),
-            food: "飲食物取扱".to_string(),
+            general_online: "オンライン一般企画".to_string(),
+            general_physical: "対面一般企画".to_string(),
+            stage_online: "オンラインステージ企画".to_string(),
+            stage_physical: "対面ステージ企画".to_string(),
+            cooking_physical: "調理".to_string(),
+            food_physical: "飲食物取扱".to_string(),
         };
         export_projects::Input {
             field_names,