+++ sos21-use-case/src/get_project_by_code.rs
@@ -67,6 +67,7 @@ where
 mod tests {
     use crate::model::project::ProjectId;
     use crate::{get_project_by_code, UseCaseError};
+    use sos21_domain::model::project::ProjectKind;
     use sos21_domain::test;
 
     // Checks that the normal user cannot read the others' project.
@@ -74,7 +75,7 @@ mod tests {
     async fn test_general_other() {
         let user = test::model::new_general_user();
         let other = test::model::new_general_user();
-        let project_other = test::model::new_general_project(other.id().clone());
+        let project_other = test::model::new_general_online_project(other.id().clone());
 
         let app = test::build_mock_app()
             .users(vec![user.clone(), other.clone()])
@@ -93,7 +94,7 @@ mod tests {
     #[tokio::test]
     async fn test_general_owner() {
         let user = test::model::new_general_user();
-        let project = test::model::new_general_project(user.id().clone());
+        let project = test::model::new_general_online_project(user.id().clone());
 
         let app = test::build_mock_app()
             .users(vec![user.clone()])
@@ -115,8 +116,10 @@ mod tests {
     async fn test_general_subowner() {
         let owner = test::model::new_general_user();
         let user = test::model::new_general_user();
-        let project =
-            test::model::new_general_project_with_subowner(owner.id().clone(), user.id().clone());
+        let project = test::model::new_general_online_project_with_subowner(
+            owner.id().clone(),
+            user.id().clone(),
+        );
 
         let app = test::build_mock_app()
             .users(vec![owner.clone(), user.clone()])
@@ -138,7 +141,7 @@ mod tests {
     async fn test_committee_other() {
         let user = test::model::new_committee_user();
         let other = test::model::new_general_user();
-        let project_other = test::model::new_general_project(other.id().clone());
+        let project_other = test::model::new_general_online_project(other.id().clone());
 
         let app = test::build_mock_app()
             .users(vec![user.clone(), other.clone()])
@@ -160,7 +163,7 @@ mod tests {
     async fn test_operator_other() {
         let user = test::model::new_operator_user();
         let other = test::model::new_general_user();
-        let project_other = test::model::new_general_project(other.id().clone());
+        let project_other = test::model::new_general_online_project(other.id().clone());
 
         let app = test::build_mock_app()
             .users(vec![user.clone(), other.clone()])
@@ -183,7 +186,7 @@ mod tests {
     async fn test_committee_nonexisting_other() {
         let user = test::model::new_committee_user();
         let other = test::model::new_general_user();
-        let project_other = test::model::new_general_project(other.id().clone());
+        let project_other = test::model::new_general_online_project(other.id().clone());
 
         let app = test::build_mock_app()
             .users(vec![user.clone(), other.clone()])
@@ -204,7 +207,7 @@ mod tests {
     async fn test_committee_different_kind_other() {
         let user = test::model::new_committee_user();
         let other = test::model::new_general_user();
-        let project_other = test::model::new_general_project(other.id().clone());
+        let project_other = test::model::new_general_online_project(other.id().clone());
 
         let app = test::build_mock_app()
             .users(vec![user.clone(), other.clone()])
@@ -214,7 +217,7 @@ mod tests {
             .await;
 
         let mut code = project_other.code();
-        code.kind.is_cooking = !code.kind.is_cooking;
+        code.kind = ProjectKind::General { is_online: false };
         assert!(matches!(
             get_project_by_code::run(&app, code.to_string()).await,
             Err(UseCaseError::UseCase(get_project_by_code::Error::NotFound))