+++ CHANGELOG.md
@@ -13,6 +13,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 ### Security
 
+## [0.6.0] - 2022-04-13
+
+### Changed
+
+- Append online `ProjectCategory` entries ([#140](https://github.com/sohosai/sos21-backend/pull/140))
+- Project Code composition
+- Switch to `sos22-backend-infrasturucture` from `sos21` one.
+
+### Removed
+
+- `affiliation` field for user ([#139](https://github.com/sohosai/sos21-backend/pull/139))
+
+### Fixed
+
+- Switch to 2022 configuration set([#138](https://github.com/sohosai/sos21-backend/pull/138))
+
 ## [0.5.3] - 2021-09-29
 
 ### Fixed